Split server tests
